station8: コンポーネント分割とProps
