+++ src/DogImage.jsx
@@ -1,7 +1,13 @@
 // @ts-check
 
-export const DogImage = () => {
-  return <></>
+export const DogImage = props => {
+  return (
+    <>
+      <div className="image">
+        <img src={props.imageUrl} />
+      </div>
+    </>
+  )
 }
 
 export default DogImage
