+++ src/Header.jsx
@@ -1,7 +1,13 @@
 // @ts-check
 
 export const Header = () => {
-  return <></>
+  return (
+    <>
+      <header className="header">
+        <h1>Dog Photo App</h1>
+      </header>
+    </>
+  )
 }
 
 export default Header
